rename files to make external use more intuitive.
