+++ lib/aws-cleaner.rb
